fix: load Tucano in bf16, not fp32 — 16x faster under memory pressure
Measured: with the gpt2-ft model (v1) also resident, fp32 Tucano pushed the
8GB CPU box into swap thrashing — one 110-token response took 553s. Switching
to bf16 (half the memory) brought the same request down to 33.8s.

Affects webapp/app.py (v2 loader) and scripts/best_of_n.py (--model tucano).

+++ scripts/best_of_n.py
@@ -43,7 +43,7 @@ def gen_tucano(prompt, n, max_new_tokens, temperature):
     from transformers import AutoModelForCausalLM, AutoTokenizer
     from peft import PeftModel
     tok = AutoTokenizer.from_pretrained(TUCANO)
-    model = AutoModelForCausalLM.from_pretrained(TUCANO, dtype=torch.float32)
+    model = AutoModelForCausalLM.from_pretrained(TUCANO, dtype=torch.bfloat16)
     if os.path.isdir(LORA):
         model = PeftModel.from_pretrained(model, LORA)
     model.eval()
